[#61] Consumer 바우처 내역 조회 추가 및 생성 Service 추가

+++ homecare/src/main/java/jaega/homecare/domain/review/dto/req/CreateReviewRequest.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public record CreateReviewRequest(
         UUID consumerId,
         UUID serviceMatchId,
-        Integer reviewScore,
+        Double reviewScore,
         String reviewContent
 ) {
 }
